REAL CHANGES DO NOT PULL IF YOU DONT WANT TO DEBUG
I edit the file for real i am not sure if it works but it is nice 
if you want the audio shuffled maybe is good time to shuffle the names in the file in order to not add extra complexity! 
shuffle the titles of the files and we keep the same indexing for both tables!